+++ src/main/java/com/MotorbikeRental/service/PaymentService.java
@@ -3,9 +3,10 @@ package com.MotorbikeRental.service;
 import org.springframework.http.ResponseEntity;
 
 import java.io.UnsupportedEncodingException;
+import java.math.BigDecimal;
 
 public interface PaymentService {
-    ResponseEntity<?> createPayment(Long id, double amount) throws UnsupportedEncodingException;
+    ResponseEntity<?> createPayment(Long id, BigDecimal amount) throws UnsupportedEncodingException;
 //    ResponseEntity<Void> returnPayment(String vnp_ResponseCode, double amount, Long id);
-    ResponseEntity<Void> returnPayment(String vnp_ResponseCode, double amount, Long id, String vnp_TxnRef);
+    ResponseEntity<Void> returnPayment(String vnp_ResponseCode, BigDecimal amount, Long id, String vnp_TxnRef);
 }
